+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [Unreleased]
+
+### Added
+
+* `--all` to show all available definitions
+
 ## [v3.4.0](https://github.com/Clay-6/WhatYouMean/releases/tag/v3.4.0)
 
 ### Added